test: split PDF visual equation component coverage
